# WIP: Validate rule options