--- a/lib/cli-engine.js
+++ b/lib/cli-engine.js
@@ -27,7 +27,8 @@ var fs = require("fs"),
     traverse = require("./util/traverse"),
     IgnoredPaths = require("./ignored-paths"),
     Config = require("./config"),
-    util = require("./util");
+    util = require("./util"),
+    validate = require("./validate-options");
 
 //------------------------------------------------------------------------------
 // Typedefs
@@ -285,6 +286,10 @@ function CLIEngine(options) {
             rules.load(rulesdir);
         });
     }
+
+    Object.keys(this.options.rules || {}).forEach(function(name) {
+        validate(name, this.options.rules[name], "CLI");
+    }.bind(this));
 }
 
 CLIEngine.prototype = {