--- a/lib/config.js
+++ b/lib/config.js
@@ -11,17 +11,19 @@
 // Requirements
 //------------------------------------------------------------------------------
 
-var fs = require("fs"),
-    path = require("path"),
-    environments = require("../conf/environments"),
-    util = require("./util"),
-    FileFinder = require("./file-finder"),
-    stripComments = require("strip-json-comments"),
-    assign = require("object-assign"),
+var assign = require("object-assign"),
     debug = require("debug"),
-    yaml = require("js-yaml"),
+    environments = require("../conf/environments"),
+    FileFinder = require("./file-finder"),
+    fs = require("fs"),
+    isAbsolutePath = require("path-is-absolute"),
+    path = require("path"),
+    rules = require('./rules'),
+    stripComments = require("strip-json-comments"),
     userHome = require("user-home"),
-    isAbsolutePath = require("path-is-absolute");
+    util = require("./util"),
+    validate = require("./validate-options"),
+    yaml = require("js-yaml");
 
 //------------------------------------------------------------------------------
 // Constants
@@ -84,6 +86,12 @@ function loadConfig(filePath) {
             }
 
             config = require(filePath);
+        }
+
+        if (config.rules) {
+            Object.keys(config.rules).forEach(function (name) {
+                validate(name, config.rules[name], filePath);
+            });
         }
 
         // If an `extends` property is defined, it represents a configuration file to use as