--- a/lib/config.js
+++ b/lib/config.js
@@ -18,7 +18,6 @@ var assign = require("object-assign"),
     fs = require("fs"),
     isAbsolutePath = require("path-is-absolute"),
     path = require("path"),
-    rules = require('./rules'),
     stripComments = require("strip-json-comments"),
     userHome = require("user-home"),
     util = require("./util"),
@@ -86,6 +85,12 @@ function loadConfig(filePath) {
             }
 
             config = require(filePath);
+        }
+
+        if (config.rules) {
+            Object.keys(config.rules).forEach(function (name) {
+                validate(name, config.rules[name], filePath);
+            });
         }
 
         if (config.rules) {