--- a/lib/config.js
+++ b/lib/config.js
@@ -93,12 +93,6 @@ function loadConfig(filePath) {
             });
         }
 
-        if (config.rules) {
-            Object.keys(config.rules).forEach(function (name) {
-                validate(name, config.rules[name], filePath);
-            });
-        }
-
         // If an `extends` property is defined, it represents a configuration file to use as
         // a "parent". Load the referenced file and merge the configuration recursively.
         if (config.extends) {