--- a/lib/eslint.js
+++ b/lib/eslint.js
@@ -623,7 +623,7 @@ module.exports = (function() {
                 var ruleCreator = rules.get(key),
                     severity = getRuleSeverity(config.rules[key]),
                     options = getRuleOptions(config.rules[key]),
-                    rule;
+                    rule, valid;
 
                 if (ruleCreator) {
                     try {