--- a/lib/eslint.js
+++ b/lib/eslint.js
@@ -8,17 +8,18 @@
 // Requirements
 //------------------------------------------------------------------------------
 
-var estraverse = require("estraverse-fb"),
-    escope = require("escope"),
-    environments = require("../conf/environments"),
-    assign = require("object-assign"),
-    rules = require("./rules"),
-    util = require("./util"),
-    RuleContext = require("./rule-context"),
-    timing = require("./timing"),
+var assign = require("object-assign"),
     createTokenStore = require("./token-store.js"),
+    environments = require("../conf/environments"),
+    escapeRegExp = require("escape-string-regexp"),
+    escope = require("escope"),
+    estraverse = require("estraverse-fb"),
     EventEmitter = require("events").EventEmitter,
-    escapeRegExp = require("escape-string-regexp");
+    RuleContext = require("./rule-context"),
+    rules = require("./rules"),
+    timing = require("./timing"),
+    util = require("./util"),
+    validate = require("./validate-options");
 
 //------------------------------------------------------------------------------
 // Helpers
@@ -238,13 +239,14 @@ function enableReporting(reportingConfig, start, rulesToEnable) {
  * Parses comments in file to extract file-specific config of rules, globals
  * and environments and merges them with global config; also code blocks
  * where reporting is disabled or enabled and merges them with reporting config.
+ * @param {string} filename The file being checked.
  * @param {ASTNode} ast The top node of the AST.
  * @param {Object} config The existing configuration data.
  * @param {Object[]} reportingConfig The existing reporting configuration data.
  * @param {Object[]} messages The messages queue.
  * @returns {void}
  */
-function modifyConfigsFromComments(ast, config, reportingConfig, messages) {
+function modifyConfigsFromComments(filename, ast, config, reportingConfig, messages) {
 
     var commentConfig = {
         astGlobals: {},
@@ -285,6 +287,7 @@ function modifyConfigsFromComments(ast, config, reportingConfig, messages) {
                         Object.keys(items).forEach(function(name) {
                             var ruleValue = items[name];
                             if (typeof ruleValue === "number" || (Array.isArray(ruleValue) && typeof ruleValue[0] === "number")) {
+                                validate(name, ruleValue, filename + " line " + comment.loc.start.line);
                                 commentRules[name] = ruleValue;
                             }
                         });
@@ -613,7 +616,7 @@ module.exports = (function() {
             currentAST = ast;
 
             // parse global comments and modify config
-            modifyConfigsFromComments(ast, config, reportingConfig, messages);
+            modifyConfigsFromComments(filename, ast, config, reportingConfig, messages);
 
             // enable appropriate rules
             Object.keys(config.rules).filter(function(key) {
@@ -623,7 +626,7 @@ module.exports = (function() {
                 var ruleCreator = rules.get(key),
                     severity = getRuleSeverity(config.rules[key]),
                     options = getRuleOptions(config.rules[key]),
-                    rule, valid;
+                    rule;
 
                 if (ruleCreator) {
                     try {