--- a/lib/rules/block-scoped-var.js
+++ b/lib/rules/block-scoped-var.js
@@ -316,3 +316,5 @@ module.exports = function(context) {
     };
 
 };
+
+module.exports.schema = [];