--- a/lib/rules/brace-style.js
+++ b/lib/rules/brace-style.js
@@ -211,3 +211,23 @@ module.exports = function(context) {
     };
 
 };
+
+module.exports.schema = {
+    "type": "array",
+    "items": [
+        {
+            "type": "string",
+            "enum": ["1tbs", "stroustrup"]
+        },
+        {
+            "type": "object",
+            "properties": {
+                "allowSingleLine": {
+                    "type": "boolean"
+                }
+            },
+            "additionalProperties": false
+        }
+    ],
+    "maxItems": 2
+};