--- a/lib/rules/brace-style.js
+++ b/lib/rules/brace-style.js
@@ -212,21 +212,17 @@ module.exports = function(context) {
 
 };
 
-module.exports.schema = {
-    "items": [
-        {
-            "type": "string",
-            "enum": ["1tbs", "stroustrup"]
+module.exports.schema = [
+    {
+        "enum": ["1tbs", "stroustrup"]
+    },
+    {
+        "type": "object",
+        "properties": {
+            "allowSingleLine": {
+                "type": "boolean"
+            }
         },
-        {
-            "type": "object",
-            "properties": {
-                "allowSingleLine": {
-                    "type": "boolean"
-                }
-            },
-            "additionalProperties": false
-        }
-    ],
-    "maxItems": 2
-};
+        "additionalProperties": false
+    }
+];