--- a/lib/rules/brace-style.js
+++ b/lib/rules/brace-style.js
@@ -213,7 +213,6 @@ module.exports = function(context) {
 };
 
 module.exports.schema = {
-    "type": "array",
     "items": [
         {
             "type": "string",