--- a/lib/rules/camelcase.js
+++ b/lib/rules/camelcase.js
@@ -97,3 +97,20 @@ module.exports = function(context) {
     };
 
 };
+
+module.exports.schema = {
+    "type": "array",
+    "items": [
+        {
+            "type": "object",
+            "properties": {
+                "properties": {
+                    "type": "string",
+                    "enum": ["always", "never"]
+                }
+            },
+            "additionalProperties": false
+        }
+    ],
+    "maxItems": 1
+};