--- a/lib/rules/camelcase.js
+++ b/lib/rules/camelcase.js
@@ -99,7 +99,6 @@ module.exports = function(context) {
 };
 
 module.exports.schema = {
-    "type": "array",
     "items": [
         {
             "type": "object",