--- a/lib/rules/camelcase.js
+++ b/lib/rules/camelcase.js
@@ -98,18 +98,14 @@ module.exports = function(context) {
 
 };
 
-module.exports.schema = {
-    "items": [
-        {
-            "type": "object",
+module.exports.schema = [
+    {
+        "type": "object",
+        "properties": {
             "properties": {
-                "properties": {
-                    "type": "string",
-                    "enum": ["always", "never"]
-                }
-            },
-            "additionalProperties": false
-        }
-    ],
-    "maxItems": 1
-};
+                "enum": ["always", "never"]
+            }
+        },
+        "additionalProperties": false
+    }
+];