--- a/lib/rules/comma-dangle.js
+++ b/lib/rules/comma-dangle.js
@@ -56,3 +56,14 @@ module.exports = function (context) {
         "ArrayExpression": checkForTrailingComma
     };
 };
+
+module.exports.schema = {
+    "type": "array",
+    "items": [
+        {
+            "type": "string",
+            "enum": ["always", "always-multiline", "never"]
+        }
+    ],
+    "maxItems": 1
+};