--- a/lib/rules/comma-dangle.js
+++ b/lib/rules/comma-dangle.js
@@ -57,12 +57,8 @@ module.exports = function (context) {
     };
 };
 
-module.exports.schema = {
-    "items": [
-        {
-            "type": "string",
-            "enum": ["always", "always-multiline", "never"]
-        }
-    ],
-    "maxItems": 1
-};
+module.exports.schema = [
+    {
+        "enum": ["always", "always-multiline", "never"]
+    }
+];