--- a/lib/rules/comma-dangle.js
+++ b/lib/rules/comma-dangle.js
@@ -58,7 +58,6 @@ module.exports = function (context) {
 };
 
 module.exports.schema = {
-    "type": "array",
     "items": [
         {
             "type": "string",