--- a/lib/rules/comma-spacing.js
+++ b/lib/rules/comma-spacing.js
@@ -157,3 +157,22 @@ module.exports = function(context) {
     };
 
 };
+
+module.exports.schema = {
+    "type": "array",
+    "items": [
+        {
+            "type": "object",
+            "properties": {
+                "before": {
+                    "type": "boolean"
+                },
+                "after": {
+                    "type": "boolean"
+                }
+            },
+            "additionalProperties": false
+        }
+    ],
+    "maxItems": 1
+};