--- a/lib/rules/comma-spacing.js
+++ b/lib/rules/comma-spacing.js
@@ -159,7 +159,6 @@ module.exports = function(context) {
 };
 
 module.exports.schema = {
-    "type": "array",
     "items": [
         {
             "type": "object",