--- a/lib/rules/comma-spacing.js
+++ b/lib/rules/comma-spacing.js
@@ -158,20 +158,17 @@ module.exports = function(context) {
 
 };
 
-module.exports.schema = {
-    "items": [
-        {
-            "type": "object",
-            "properties": {
-                "before": {
-                    "type": "boolean"
-                },
-                "after": {
-                    "type": "boolean"
-                }
+module.exports.schema = [
+    {
+        "type": "object",
+        "properties": {
+            "before": {
+                "type": "boolean"
             },
-            "additionalProperties": false
-        }
-    ],
-    "maxItems": 1
-};
+            "after": {
+                "type": "boolean"
+            }
+        },
+        "additionalProperties": false
+    }
+];