--- a/lib/rules/comma-style.js
+++ b/lib/rules/comma-style.js
@@ -175,3 +175,9 @@ module.exports = function(context) {
 
     return nodes;
 };
+
+module.exports.schema = [
+    {
+        "enum": ["first", "last"]
+    }
+];