--- a/lib/rules/comma-style.js
+++ b/lib/rules/comma-style.js
@@ -179,5 +179,17 @@ module.exports = function(context) {
 module.exports.schema = [
     {
         "enum": ["first", "last"]
+    },
+    {
+        "type": "object",
+        "properties": {
+            "exceptions": {
+                "type": "object",
+                "additionalProperties": {
+                    "type": "boolean"
+                }
+            }
+        },
+        "additionalProperties": false
     }
 ];