--- a/lib/rules/complexity.js
+++ b/lib/rules/complexity.js
@@ -86,3 +86,9 @@ module.exports = function(context) {
     };
 
 };
+
+module.exports.schema = [
+    {
+        "type": "integer"
+    }
+];