--- a/lib/rules/consistent-return.js
+++ b/lib/rules/consistent-return.js
@@ -71,3 +71,5 @@ module.exports = function(context) {
     };
 
 };
+
+module.exports.schema = [];