--- a/lib/rules/consistent-this.js
+++ b/lib/rules/consistent-this.js
@@ -106,3 +106,9 @@ module.exports = function(context) {
     };
 
 };
+
+module.exports.schema = [
+    {
+        "type": "string"
+    }
+];