--- a/lib/rules/curly.js
+++ b/lib/rules/curly.js
@@ -101,3 +101,14 @@ module.exports = function(context) {
     };
 
 };
+
+module.exports.schema = {
+    "type": "array",
+    "items": [
+        {
+            "type": "string",
+            "enum": ["all", "multi", "multi-line"]
+        }
+    ],
+    "maxItems": 1
+};