--- a/lib/rules/curly.js
+++ b/lib/rules/curly.js
@@ -103,7 +103,6 @@ module.exports = function(context) {
 };
 
 module.exports.schema = {
-    "type": "array",
     "items": [
         {
             "type": "string",