--- a/lib/rules/curly.js
+++ b/lib/rules/curly.js
@@ -102,12 +102,8 @@ module.exports = function(context) {
 
 };
 
-module.exports.schema = {
-    "items": [
-        {
-            "type": "string",
-            "enum": ["all", "multi", "multi-line"]
-        }
-    ],
-    "maxItems": 1
-};
+module.exports.schema = [
+    {
+        "enum": ["all", "multi", "multi-line"]
+    }
+];