--- a/lib/rules/default-case.js
+++ b/lib/rules/default-case.js
@@ -62,3 +62,5 @@ module.exports = function(context) {
         }
     };
 };
+
+module.exports.schema = [];