--- a/lib/rules/dot-notation.js
+++ b/lib/rules/dot-notation.js
@@ -102,3 +102,22 @@ module.exports = function(context) {
         }
     };
 };
+
+module.exports.schema = {
+    "type": "array",
+    "items": [
+        {
+            "type": "object",
+            "properties": {
+                "allowKeywords": {
+                    "type": "boolean"
+                },
+                "allowPattern": {
+                    "type": "string"
+                }
+            },
+            "additionalProperties": false
+        }
+    ],
+    "maxItems": 1
+};