--- a/lib/rules/dot-notation.js
+++ b/lib/rules/dot-notation.js
@@ -104,7 +104,6 @@ module.exports = function(context) {
 };
 
 module.exports.schema = {
-    "type": "array",
     "items": [
         {
             "type": "object",