--- a/lib/rules/dot-notation.js
+++ b/lib/rules/dot-notation.js
@@ -103,20 +103,17 @@ module.exports = function(context) {
     };
 };
 
-module.exports.schema = {
-    "items": [
-        {
-            "type": "object",
-            "properties": {
-                "allowKeywords": {
-                    "type": "boolean"
-                },
-                "allowPattern": {
-                    "type": "string"
-                }
+module.exports.schema = [
+    {
+        "type": "object",
+        "properties": {
+            "allowKeywords": {
+                "type": "boolean"
             },
-            "additionalProperties": false
-        }
-    ],
-    "maxItems": 1
-};
+            "allowPattern": {
+                "type": "string"
+            }
+        },
+        "additionalProperties": false
+    }
+];