--- a/lib/rules/eol-last.js
+++ b/lib/rules/eol-last.js
@@ -34,3 +34,5 @@ module.exports = function(context) {
     };
 
 };
+
+module.exports.schema = [];