--- a/lib/rules/eqeqeq.js
+++ b/lib/rules/eqeqeq.js
@@ -88,3 +88,14 @@ module.exports = function(context) {
     };
 
 };
+
+module.exports.schema = {
+    "type": "array",
+    "items": [
+        {
+            "type": "string",
+            "enum": ["smart", "allow-null"]
+        }
+    ],
+    "maxItems": 1
+};