--- a/lib/rules/eqeqeq.js
+++ b/lib/rules/eqeqeq.js
@@ -90,7 +90,6 @@ module.exports = function(context) {
 };
 
 module.exports.schema = {
-    "type": "array",
     "items": [
         {
             "type": "string",