--- a/lib/rules/eqeqeq.js
+++ b/lib/rules/eqeqeq.js
@@ -89,12 +89,8 @@ module.exports = function(context) {
 
 };
 
-module.exports.schema = {
-    "items": [
-        {
-            "type": "string",
-            "enum": ["smart", "allow-null"]
-        }
-    ],
-    "maxItems": 1
-};
+module.exports.schema = [
+    {
+        "enum": ["smart", "allow-null"]
+    }
+];