--- a/lib/rules/func-names.js
+++ b/lib/rules/func-names.js
@@ -41,3 +41,5 @@ module.exports = function(context) {
         }
     };
 };
+
+module.exports.schema = [];