--- a/lib/rules/func-style.js
+++ b/lib/rules/func-style.js
@@ -41,3 +41,9 @@ module.exports = function(context) {
     };
 
 };
+
+module.exports.schema = [
+    {
+        "enum": ["declaration", "expression"]
+    }
+];