--- a/lib/rules/generator-star-spacing.js
+++ b/lib/rules/generator-star-spacing.js
@@ -79,3 +79,5 @@ module.exports = function(context) {
     };
 
 };
+
+module.exports.schema = [];