--- a/lib/rules/generator-star.js
+++ b/lib/rules/generator-star.js
@@ -68,3 +68,5 @@ module.exports = function(context) {
     };
 
 };
+
+module.exports.schema = [];