--- a/lib/rules/generator-star.js
+++ b/lib/rules/generator-star.js
@@ -69,4 +69,8 @@ module.exports = function(context) {
 
 };
 
-module.exports.schema = [];
+module.exports.schema = [
+    {
+        "enum": ["start", "middle", "end"]
+    }
+];