--- a/lib/rules/global-strict.js
+++ b/lib/rules/global-strict.js
@@ -41,3 +41,9 @@ module.exports = function(context) {
     }
 
 };
+
+module.exports.schema = [
+    {
+        "enum": ["always", "never"]
+    }
+];