--- a/lib/rules/guard-for-in.js
+++ b/lib/rules/guard-for-in.js
@@ -28,3 +28,5 @@ module.exports = function(context) {
     };
 
 };
+
+module.exports.schema = [];