--- a/lib/rules/handle-callback-err.js
+++ b/lib/rules/handle-callback-err.js
@@ -116,3 +116,5 @@ module.exports = function(context) {
     };
 
 };
+
+module.exports.schema = [];