--- a/lib/rules/indent.js
+++ b/lib/rules/indent.js
@@ -462,3 +462,5 @@ module.exports = function (context) {
     };
 
 };
+
+module.exports.schema = [];