--- a/lib/rules/indent.js
+++ b/lib/rules/indent.js
@@ -463,4 +463,24 @@ module.exports = function (context) {
 
 };
 
-module.exports.schema = [];
+module.exports.schema = [
+    {
+        "oneOf": [
+            {
+                "enum": ["tab"]
+            },
+            {
+                "type": "integer"
+            }
+        ]
+    },
+    {
+        "type": "object",
+        "properties": {
+            "indentSwitchCase": {
+                "type": "boolean"
+            }
+        },
+        "additionalProperties": false
+    }
+];