--- a/lib/rules/key-spacing.js
+++ b/lib/rules/key-spacing.js
@@ -305,3 +305,26 @@ module.exports = function(context) {
     }
 
 };
+
+module.exports.schema = {
+    "type": "array",
+    "items": [
+        {
+            "type": "object",
+            "properties": {
+                "align": {
+                    "type": "string",
+                    "enum": ["colon", "value"]
+                },
+                "beforeColon": {
+                    "type": "boolean"
+                },
+                "afterColon": {
+                    "type": "boolean"
+                }
+            },
+            "additionalProperties": false
+        }
+    ],
+    "maxItems": 1
+};