--- a/lib/rules/key-spacing.js
+++ b/lib/rules/key-spacing.js
@@ -306,24 +306,20 @@ module.exports = function(context) {
 
 };
 
-module.exports.schema = {
-    "items": [
-        {
-            "type": "object",
-            "properties": {
-                "align": {
-                    "type": "string",
-                    "enum": ["colon", "value"]
-                },
-                "beforeColon": {
-                    "type": "boolean"
-                },
-                "afterColon": {
-                    "type": "boolean"
-                }
+module.exports.schema = [
+    {
+        "type": "object",
+        "properties": {
+            "align": {
+                "enum": ["colon", "value"]
             },
-            "additionalProperties": false
-        }
-    ],
-    "maxItems": 1
-};
+            "beforeColon": {
+                "type": "boolean"
+            },
+            "afterColon": {
+                "type": "boolean"
+            }
+        },
+        "additionalProperties": false
+    }
+];