--- a/lib/rules/key-spacing.js
+++ b/lib/rules/key-spacing.js
@@ -307,7 +307,6 @@ module.exports = function(context) {
 };
 
 module.exports.schema = {
-    "type": "array",
     "items": [
         {
             "type": "object",