--- a/lib/rules/linebreak-style.js
+++ b/lib/rules/linebreak-style.js
@@ -30,3 +30,9 @@ module.exports = function (context) {
         }
     };
 };
+
+module.exports.schema = [
+    {
+        "enum": ["unix", "windows"]
+    }
+];