--- a/lib/rules/max-depth.js
+++ b/lib/rules/max-depth.js
@@ -81,3 +81,9 @@ module.exports = function(context) {
     };
 
 };
+
+module.exports.schema = [
+    {
+        "type": "integer"
+    }
+];