--- a/lib/rules/max-len.js
+++ b/lib/rules/max-len.js
@@ -63,3 +63,19 @@ module.exports = function(context) {
     };
 
 };
+
+module.exports.schema = {
+    "type": "array",
+    "items": [
+        {
+            "type": "integer",
+            "minimum": 0
+        },
+        {
+            "type": "integer",
+            "minimum": 0
+        }
+    ],
+    "minItems": 2,
+    "maxItems": 2
+};