--- a/lib/rules/max-len.js
+++ b/lib/rules/max-len.js
@@ -65,7 +65,6 @@ module.exports = function(context) {
 };
 
 module.exports.schema = {
-    "type": "array",
     "items": [
         {
             "type": "integer",