--- a/lib/rules/max-len.js
+++ b/lib/rules/max-len.js
@@ -32,9 +32,9 @@ module.exports = function(context) {
         return result;
     }
 
-    var tabWidth = context.options[1];
+    var tabWidth = context.options[1] || 4; // TODO: Add tests for default
 
-    var maxLength = context.options[0],
+    var maxLength = context.options[0] || 80,
         tabString = stringRepeat(" ", tabWidth);
 
     //--------------------------------------------------------------------------
@@ -64,17 +64,13 @@ module.exports = function(context) {
 
 };
 
-module.exports.schema = {
-    "items": [
-        {
-            "type": "integer",
-            "minimum": 0
-        },
-        {
-            "type": "integer",
-            "minimum": 0
-        }
-    ],
-    "minItems": 2,
-    "maxItems": 2
-};
+module.exports.schema = [
+    {
+        "type": "integer",
+        "minimum": 0
+    },
+    {
+        "type": "integer",
+        "minimum": 0
+    }
+];