--- a/lib/rules/max-len.js
+++ b/lib/rules/max-len.js
@@ -32,7 +32,7 @@ module.exports = function(context) {
         return result;
     }
 
-    var tabWidth = context.options[1] || 4; // TODO: Add tests for default
+    var tabWidth = context.options[1] || 4;
 
     var maxLength = context.options[0] || 80,
         tabString = stringRepeat(" ", tabWidth);