--- a/lib/rules/max-params.js
+++ b/lib/rules/max-params.js
@@ -37,3 +37,9 @@ module.exports = function(context) {
     };
 
 };
+
+module.exports.schema = [
+    {
+        "type": "integer"
+    }
+];