--- a/lib/rules/max-statements.js
+++ b/lib/rules/max-statements.js
@@ -53,3 +53,9 @@ module.exports = function(context) {
     };
 
 };
+
+module.exports.schema = [
+    {
+        "type": "integer"
+    }
+];