--- a/lib/rules/new-cap.js
+++ b/lib/rules/new-cap.js
@@ -195,3 +195,34 @@ module.exports = function(context) {
 
     return listeners;
 };
+
+module.exports.schema = {
+    "type": "array",
+    "items": [
+        {
+            "type": "object",
+            "properties": {
+                "newIsCap": {
+                    "type": "boolean"
+                },
+                "capIsNew": {
+                    "type": "boolean"
+                },
+                "newIsCapExceptions": {
+                    "type": "array",
+                    "items": {
+                        "type": "string"
+                    }
+                },
+                "capIsNewExceptions": {
+                    "type": "array",
+                    "items": {
+                        "type": "string"
+                    }
+                }
+            },
+            "additionalProperties": false
+        }
+    ],
+    "maxItems": 1
+};