--- a/lib/rules/new-cap.js
+++ b/lib/rules/new-cap.js
@@ -196,32 +196,29 @@ module.exports = function(context) {
     return listeners;
 };
 
-module.exports.schema = {
-    "items": [
-        {
-            "type": "object",
-            "properties": {
-                "newIsCap": {
-                    "type": "boolean"
-                },
-                "capIsNew": {
-                    "type": "boolean"
-                },
-                "newIsCapExceptions": {
-                    "type": "array",
-                    "items": {
-                        "type": "string"
-                    }
-                },
-                "capIsNewExceptions": {
-                    "type": "array",
-                    "items": {
-                        "type": "string"
-                    }
+module.exports.schema = [
+    {
+        "type": "object",
+        "properties": {
+            "newIsCap": {
+                "type": "boolean"
+            },
+            "capIsNew": {
+                "type": "boolean"
+            },
+            "newIsCapExceptions": {
+                "type": "array",
+                "items": {
+                    "type": "string"
                 }
             },
-            "additionalProperties": false
-        }
-    ],
-    "maxItems": 1
-};
+            "capIsNewExceptions": {
+                "type": "array",
+                "items": {
+                    "type": "string"
+                }
+            }
+        },
+        "additionalProperties": false
+    }
+];