--- a/lib/rules/new-cap.js
+++ b/lib/rules/new-cap.js
@@ -197,7 +197,6 @@ module.exports = function(context) {
 };
 
 module.exports.schema = {
-    "type": "array",
     "items": [
         {
             "type": "object",