--- a/lib/rules/new-parens.js
+++ b/lib/rules/new-parens.js
@@ -25,3 +25,5 @@ module.exports = function(context) {
     };
 
 };
+
+module.exports.schema = [];