--- a/lib/rules/newline-after-var.js
+++ b/lib/rules/newline-after-var.js
@@ -120,4 +120,8 @@ module.exports = function(context) {
 
 };
 
-module.exports.schema = [];
+module.exports.schema = [
+    {
+        "enum": ["never", "always"]
+    }
+];