--- a/lib/rules/newline-after-var.js
+++ b/lib/rules/newline-after-var.js
@@ -119,3 +119,5 @@ module.exports = function(context) {
     };
 
 };
+
+module.exports.schema = [];