--- a/lib/rules/no-alert.js
+++ b/lib/rules/no-alert.js
@@ -52,3 +52,5 @@ module.exports = function(context) {
     };
 
 };
+
+module.exports.schema = [];