--- a/lib/rules/no-array-constructor.js
+++ b/lib/rules/no-array-constructor.js
@@ -27,3 +27,5 @@ module.exports = function(context) {
     };
 
 };
+
+module.exports.schema = [];