--- a/lib/rules/no-bitwise.js
+++ b/lib/rules/no-bitwise.js
@@ -53,3 +53,5 @@ module.exports = function(context) {
     };
 
 };
+
+module.exports.schema = [];