--- a/lib/rules/no-caller.js
+++ b/lib/rules/no-caller.js
@@ -25,3 +25,5 @@ module.exports = function(context) {
     };
 
 };
+
+module.exports.schema = [];