--- a/lib/rules/no-catch-shadow.js
+++ b/lib/rules/no-catch-shadow.js
@@ -48,3 +48,5 @@ module.exports = function(context) {
     };
 
 };
+
+module.exports.schema = [];