--- a/lib/rules/no-comma-dangle.js
+++ b/lib/rules/no-comma-dangle.js
@@ -41,3 +41,5 @@ module.exports = function(context) {
     };
 
 };
+
+module.exports.schema = [];