--- a/lib/rules/no-cond-assign.js
+++ b/lib/rules/no-cond-assign.js
@@ -115,3 +115,14 @@ module.exports = function(context) {
     };
 
 };
+
+module.exports.schema = {
+    "type": "array",
+    "items": [
+        {
+            "type": "string",
+            "enum": ["except-parens", "always"]
+        }
+    ],
+    "maxItems": 1
+};