--- a/lib/rules/no-cond-assign.js
+++ b/lib/rules/no-cond-assign.js
@@ -116,12 +116,8 @@ module.exports = function(context) {
 
 };
 
-module.exports.schema = {
-    "items": [
-        {
-            "type": "string",
-            "enum": ["except-parens", "always"]
-        }
-    ],
-    "maxItems": 1
-};
+module.exports.schema = [
+    {
+        "enum": ["except-parens", "always"]
+    }
+];