--- a/lib/rules/no-console.js
+++ b/lib/rules/no-console.js
@@ -23,3 +23,5 @@ module.exports = function(context) {
     };
 
 };
+
+module.exports.schema = [];