--- a/lib/rules/no-constant-condition.js
+++ b/lib/rules/no-constant-condition.js
@@ -69,3 +69,5 @@ module.exports = function(context) {
     };
 
 };
+
+module.exports.schema = [];