--- a/lib/rules/no-continue.js
+++ b/lib/rules/no-continue.js
@@ -19,3 +19,5 @@ module.exports = function(context) {
     };
 
 };
+
+module.exports.schema = [];