--- a/lib/rules/no-control-regex.js
+++ b/lib/rules/no-control-regex.js
@@ -54,3 +54,5 @@ module.exports = function(context) {
     };
 
 };
+
+module.exports.schema = [];