--- a/lib/rules/no-debugger.js
+++ b/lib/rules/no-debugger.js
@@ -18,3 +18,5 @@ module.exports = function(context) {
     };
 
 };
+
+module.exports.schema = [];