--- a/lib/rules/no-delete-var.js
+++ b/lib/rules/no-delete-var.js
@@ -21,3 +21,5 @@ module.exports = function(context) {
     };
 
 };
+
+module.exports.schema = [];