--- a/lib/rules/no-div-regex.js
+++ b/lib/rules/no-div-regex.js
@@ -23,3 +23,5 @@ module.exports = function(context) {
     };
 
 };
+
+module.exports.schema = [];