--- a/lib/rules/no-dupe-args.js
+++ b/lib/rules/no-dupe-args.js
@@ -81,3 +81,5 @@ module.exports = function(context) {
     };
 
 };
+
+module.exports.schema = [];