--- a/lib/rules/no-dupe-keys.js
+++ b/lib/rules/no-dupe-keys.js
@@ -39,3 +39,5 @@ module.exports = function(context) {
     };
 
 };
+
+module.exports.schema = [];