--- a/lib/rules/no-duplicate-case.js
+++ b/lib/rules/no-duplicate-case.js
@@ -57,3 +57,5 @@ module.exports = function(context) {
     };
 
 };
+
+module.exports.schema = [];