--- a/lib/rules/no-else-return.js
+++ b/lib/rules/no-else-return.js
@@ -121,3 +121,5 @@ module.exports = function(context) {
     };
 
 };
+
+module.exports.schema = [];