--- a/lib/rules/no-empty-class.js
+++ b/lib/rules/no-empty-class.js
@@ -41,3 +41,5 @@ module.exports = function(context) {
     };
 
 };
+
+module.exports.schema = [];