This is a code-included, tests-passing request for bikeshedding that might address the discussions in #967 and (eventually) #1568. This is not ready for a line-by-line code review until higher-level concepts have been addressed and solidified.

With this approach, validation is optional. Rules opt in by exporting a `schema` object describing a JSON schema for the options the rule accepts. Most rules have very similar option formats, so this wasn't difficult at all.

### Example

```js
// test.js
var foo = 4 == (2 * 2);
```

```js
// config.json
{
    "rules": {
        "eqeqeq": [2, "samrt"]
    }
}
```

```sh
$ eslint --reset --no-eslintrc -c config.json test.js

test.js
  0:0  error  Invalid configuration: ["samrt"] must be an enum value  eqeqeq

✖ 1 problem (1 error, 0 warnings)
```

### Performance

So far, I've created schemas for all rules that are enabled by default in order to get a good picture of the performance impact, which, happily, turned out to be minor.

`master`:

```sh
CPU Speed is 2200 with multiplier 7500000
Performance Run #1:  1994.5480929999999ms
Performance Run #2:  2227.983234ms
Performance Run #3:  2227.137346ms
Performance Run #4:  2177.492682ms
Performance Run #5:  2025.890174ms
Performance budget ok:  2177.492682ms (limit: 3409.090909090909ms)
```

This PR:

```sh
CPU Speed is 2200 with multiplier 7500000
Performance Run #1:  2228.800724ms
Performance Run #2:  2247.415942ms
Performance Run #3:  2255.482898ms
Performance Run #4:  2244.724581ms
Performance Run #5:  2235.863643ms
Performance budget ok:  2244.724581ms (limit: 3409.090909090909ms)
```

Note that these performance numbers are for linting a single file. When linting a whole directory, part of the validation cost is amortized over all files because [is-my-json-valid](https://www.npmjs.com/package/is-my-json-valid) uses code generation to create the validators, which are generated once, cached, and reused.

### Configuration guidelines

These are some principles that we seem to prefer, but I'd like to consider formalizing them. These are just ones I've noticed or think make sense; feedback is requested.

1. All rules should have sensible defaults, such that they can be enabled and will do *something* without having to specify any options.

    **Example**: `"quotes": 2` can assume double quotes since neither mode was specified.

1. A rule whose primary option is a mode may specify that mode using a descriptive string as the first option.

    **Example**: `"quotes": [2, "single"]` intuitively should enforce single quotes.

1. All other (non-mode) options should be named properties of a single options object. This helps avoid confusion as to the order of parameters, and it allows some to be left as default while only specifying deviations from the defaults.

    **Example**: `"semi-spacing": [2, { "before": false, "after": true }]`

1. If we have validation, out of range values should fail validation. Previously, some rules, `comma-dangle` being one, considered unknown values to mean the default.

    **Example**: `"quotes": [2, "snigle"]` should fail, since clearly `"single"` was intended, yet assuming the default of `"double"` would lead to confusing behavior until the error was noticed. (This is a trivial example. Other, deeper options fail silently and are much harder to debug.)

Some rules still support for deprecated options. For example, `no-mixed-spaces-and-tabs` also accepts a boolean value, though it's not documented, and a source code comment suggests adding a deprecation warning. Since this would be a breaking change, it could also be a good time to remove support for deprecated options concurrently, though I've not done so here.

Thoughts?

### Feedback requested

- General bikeshedding: is this the right approach?
- What should error messages look like? I'm not attached to what exists currently; I think they could be more helpful.
- Simplifying schemas: since there's so much in common, it might be feasible to offload some of the schema-creation work to the validator.
- Are there guiding principles anywhere for how rules should be configured? I've sketched out some ideas above; are they reasonable?