@ilyavolodin 2015-03-30
In general approach looks pretty good to me. I would make a few changes, like, adding possible error message into schema as well as exposing a function on cli to dump all schemas for all rules, but all of that can be added later as an independent commit (this would allow us to automatically improve rules documentation as well as allow editors to get all possible configurations for rules).
My only question is, should we validate configuration as part of a normal run, or just have a command line option that will validate configuration, without actually validating code?