@ilyavolodin 2015-03-30
Also, one more thing, I'm not sure that would be possible, but it would be nice if the error message would point to the configuration file with line number (don't think it is, since it's inside the loop that processes files, so it will always point to the current file that is being validated).