@nzakas 2015-03-30
I think the idea is correct, just not sure about the implementation. The schema looks a bit complicated, which means we would need a schema validator, and that starts to get hairy. I'm wondering if we should use something like [joi](https://github.com/hapijs/joi) instead. 

@ilyavolodin I definitely don't want a separate validation step, it should be fine as part of the normal run, otherwise people will rarely be aware of errors.

I'm not sure about the error representation either. The error could be in an eslintrc file or a JS file due to inline comment. We either need to be better about reporting where the error occurred or just throw the error and let ESLint crash.