@btmills 2015-04-07
> it would be nice if the error message would point to the configuration file with line number

```js
// test.js
/* eslint quotes: [3, "double"] */
var answer = 6 * 7;
```

```js
// config.json
{
    "rules": {
        "semi": [2, "laways"]
    }
}
```

```sh
$ eslint test.js
... error thrown

Error: test.js line 2:
	Configuration for rule "quotes" is invalid:
	Error level must be between 0 (off) and 2 (warning).

... stack trace
```

```sh
$ eslint -c config.json test.js
... error thrown

Error: /Users/brandon/code/eslint/eslint/config.json:
	Configuration for rule "semi" is invalid:
	Value "laways" must be an enum value.

... stack trace
```

```sh
$ eslint --rule 'eqeqeq: [2, "smart", "asdf"]' test.js
... error thrown

Error: CLI:
	Configuration for rule "eqeqeq" is invalid:
	Value "2,smart,asdf" has more items than allowed.

... stack trace
```

> The schema looks a bit complicated... I'm wondering if we should use something like joi instead.

I looked into joi, but I found no way to programmatically inspect schemas. Right now we only need validation, but getting into configuration setup and fuzz testing, that becomes more important. I did, however, offload some of the schema creation onto the validator. Half of the schemas now become just an enum, `semi` being one example:

```js
module.exports.schema = [
    {
        "enum": ["always", "never"]
    }
];
```

As for Travis, it's failing because of code coverage; I didn't break any logic.