@nzakas 2015-04-08
Here's my concern: representing data as JSON is a solved problem. If we're creating a new schema type, then we're also on the hook for creating something to validate that schema. I'd much prefer using an existing solution vs. creating something completely new.