@ilyavolodin 2015-04-08
Just in case you change the format of schema to something else, please keep in mind I would like to add the following information to the schema later at some point. Here's the information that I would like to add:
```json
{
"description": "very short description of the rule",
"errors": ["list of all of the errors that can be reported by this rule"],
"options": { "always": "description", "never": "description" }
```
Or something like that. This would allow us to enhance our documentation automatically, as well as provide a way to extract all of the options for all rules for integrations and things like web .eslintrc builder.