@nzakas 2015-04-08
@ilyavolodin we can always add extra properties to the rules themselves, this doesn't have to be part of the schema