@doberkofler 2015-04-11
@btmills I love the idea to have the rules validated and it's great what you have prepared.
I might be far off but why could not just each rule (maybe also optionally) implement a callback that is responsible to validate it's own configuration and to report any errors.
I see the advantages in a standardized schema but it also seems a little complicated and it still requires the schema's themselves to be validated as @nzakas points out.