@btmills 2015-04-15
> I'd much prefer using an existing solution vs. creating something completely new.

@nzakas This is still using unmodified JSON schema. However, since much of each rule's schema was redundant, I had the rules just return the `items` in the options array, and `makeSchema` wraps that, inserting the reporting level as the first item and specifying `type`, `minItems`, `maxItems`. A rule's `schema` property is an array of individual JSON schemas for each option, and the result of `makeSchema` is a full JSON schema for the rule's options.

For example, `quotes` exports the following array, each element of which is a valid JSON schema by itself:

```json
[
    {
        "enum": ["single", "double"]
    },
    {
        "enum": ["avoid-escape"]
    }
]
```

Wrapping it with `makeSchema` produces the full JSON schema for all of the rule's options (ignore the `title` and `description` attributes for now):

```json
{
    "title": "quotes",
    "description": "Specify whether double or single quotes should be used.",
    "type": "array",
    "items": [
        {
            "enum": [0, 1, 2],
            "description": "Reporting level - 0 for off, 1 for warning, or 2 for error."
        },
        {
            "enum": ["single", "double"],
            "description": "Enforce single or double quotes."
        },
        {
            "enum": ["avoid-escape"],
            "description": "Avoid escaped quotes inside strings by allowing both quote types when necessary."
        }
    ],
    "minItems": 1,
    "maxItems": 3
}
```

Because there are valid schemas in both phases, we could validate them against JSON schema's schema in either place.

> I would like to add the following information to the schema later at some point.

@ilyavolodin I annotated the above example with `title` and `description` attributes, both of which are purely informative and part of the JSON schema spec. Would those be sufficient for what you're imagining? (The overall `description` I pulled straight from the `docs/rules` index. I could see it being exported by the rule as a `description` property to be used in a configuration helper akin to #2302. It would also enable us to generate `docs/rules/README.md` programmatically. The list of possible error messages might be exported similarly.)

> why could not just each rule (maybe also optionally) implement a callback

@doberkofler Callbacks were actually the first thing I considered way back when I started working on fuzz testing, see [#1568 (comment)](https://github.com/eslint/eslint/issues/1568#issuecomment-66905928). However, as pointed out by @nzakas, fuzz testing and config validation would require two separate callbacks. We'd need another for some sort of autoconfiguration utility. That's why we need some standard data format that facilitates programmatic inspection and generation. From what I've seen, JSON schema seems to be that.