@ilyavolodin 2015-04-15
@btmills Yes, that would work perfectly for auto configuration and auto documentation. One last thing that would be nice to add is some sort of way to mark default options (if there are any).